Record both address bounds of every unpacked dimension on a variable and a port, and address an array port and a multidimensional array from them
An unpacked array declared as a module port, and one of more than one dimension, could not be synthesized at all: the reader was told "an element of this unpacked array has no lowering in the synthesizer" against §11.5.2, which names a subclause rather than anything the design got wrong. A port dimension written over a parameter was worse than that. It was folded with no scope, so `input logic [7:0] mem [N]` recorded no dimension, RtlirPort::num_unpacked_dims came out zero, the synthesizer was never told the name is an array, and `mem[2]` lowered as a §11.5.1 bit-select of one 8-bit element and exited 0.

§7.4.2 (printed page 154) rules that an unpacked array is "declared by specifying the element address range(s) after the declared identifier", that a dimension's "first value may be greater than, equal to, or less than the second value", that a bound "may be any integer value -- positive, negative, or zero", and that "[size] shall mean the same as [0:size-1]". §11.5.2 (printed page 297) rules that "the address bounds given in the declaration of the memory determine the effect of the address expression", that an invalid address goes to §7.4.5, and that "the desired word shall first be selected by supplying an address for each dimension". So both bounds of every dimension are what an address is resolved against, and the dimension count is what says how many addresses a reference owes.

RtlirUnpackedDim is that pair, and RtlirVariable and RtlirPort now carry one per dimension that folded, in declaration order, beside a count of the dimensions the declaration wrote. The count is what tells `logic [7:0] m [1:4]` from `logic [7:0] m [1:4][]`, which recorded the same thing before and let the second be read as an array of words. The vector is filled for one dimension as well as for two, so a consumer reads one family of fields whatever the declaration. RtlirVariable::unpacked_lo widens to int64_t, since §7.4.2 admits a negative bound and `int x [-3:5]` recorded 4294967293.

The port path folds in the port's own parameter scope, which Elaborator::ElaboratePorts already builds and ElaborateOnePort already passes to EvalTypeWidth on the line above the call this changes. A dimension that still does not fold is now reported against §7.4.2 rather than skipped, since a skipped dimension is what made the miscount silent.

SynthLower::ArrayShape carries a low bound and an element count per dimension, SynthLower::RecordArrayShape accepts any dimension count and refuses only a declaration whose recorded dimensions fall short of the count it wrote, and SynthLower::MapPorts calls it for a port as well as a variable so an array port gets one element's storage per address rather than one element's in total. SynthLower::ResolveSelect classifies a chain of selects by how many addresses it carries against how many dimensions the array declares: exactly as many names an element, fewer names a slice this does not name, and more is a §11.5.1 select within the element, which is the recursion that was already there. An address outside a dimension's bounds names a run past the whole storage rather than the offset the arithmetic gives, because an address out of one dimension's bounds otherwise lands inside another dimension's elements.

The simulator keeps reading the smaller bound of each pair, which is what ArrayInfo holds; its dim_los is uint32_t, so a negative bound still does not survive that copy, and carrying one into the simulator is separate work.

The nineteen tests are written so that no dimension starts at zero except where the `[size]` form is what is under test, because where a dimension starts at zero an address and an element offset are the same number and a case on it passes whether the bound was recorded or not. The seven synthesizer cases drive the netlist through ExpectInputSweep rather than asserting a graph came back.

Closes #3047

+++ src/elaborator/elaborator_decls.cpp
@@ -68,7 +68,7 @@ static bool TryParseRangeDim(const Expr* dim, RtlirVariable& var,
   if (!lval || !rval) return false;
   auto lo = std::min(*lval, *rval);
   auto hi = std::max(*lval, *rval);
-  var.unpacked_lo = static_cast<uint32_t>(lo);
+  var.unpacked_lo = lo;
   var.unpacked_size = static_cast<uint32_t>(hi - lo + 1);
   var.is_descending = (*lval > *rval);
   return true;
